Refresh token rotation

+++ controllers/adminController.js
@@ -1,21 +1,25 @@
 /* eslint-disable consistent-return */
+const { USER_STATUS } = require('../config/token&CommonVar');
 const User = require('../models/User.model');
-const { sendResponse, toggleUserStatus } = require('../utils/commonMethod');
+const {
+  sendResponse, handleUserStatusblock, handleUserStatusUnblock,
+} = require('../utils/commonMethod');
 
 const handleAdminPanel = async (req, res, next) => {
   try {
-    const userId = req.user;
+    // check user id from jwt
+    const userId = req?.user;
     const validUser = await User.findOne({ _id: userId });
 
     if (!validUser) {
       return sendResponse(res, 404, 'User not found');
     }
 
-    if (validUser.status === 'blocked') {
-      return sendResponse(res, 404, 'User is blocked');
+    if (validUser.status === USER_STATUS.Blocked) {
+      return sendResponse(res, 404, 'Blocked by Admin');
     }
 
-    sendResponse(res, 200, 'Admin Panel accessed successfully');
+    return sendResponse(res, 200, 'Admin Panel accessed successfully');
   } catch (err) {
     next(err);
   }
@@ -24,31 +28,51 @@ const handleAdminPanel = async (req, res, next) => {
 const getUsers = async (req, res, next) => {
   try {
     const foundUsers = await User.find().select('-password -refreshToken -roles');
-    sendResponse(res, 200, 'All users fetched successfully', foundUsers);
+
+    if (!foundUsers.length) {
+      return sendResponse(res, 204, 'No users found');
+    }
+
+    return sendResponse(res, 200, 'All users fetched successfully', foundUsers);
   } catch (err) {
     next(err);
   }
 };
 
-const updateUser = async (req, res, next) => {
+const updateUsersBlock = async (req, res, next) => {
   try {
-    const { userId } = req.params;
-    await toggleUserStatus(userId);
+    if (!req?.body?.userIds) {
+      return sendResponse(res, 400, 'User IDs required');
+    }
 
-    sendResponse(res, 200, 'User status toggled successfully');
-  } catch (err) {
-    next(err);
-  }
-};
-
-const updateUsers = async (req, res, next) => {
-  try {
     const { userIds } = req.body;
 
-    // parallelize the asynchronous operations
-    await Promise.all(userIds.map((userId) => toggleUserStatus(userId)));
+    try {
+      // parallelize the asynchronous operations
+      await Promise.all(userIds.map((userId) => handleUserStatusblock(userId)));
+      return sendResponse(res, 200, 'status blocked successfully');
+    } catch (dbError) {
+      sendResponse(res, 400, String(dbError));
+    }
+  } catch (err) {
+    next(err);
+  }
+};
+const updateUsersUnblock = async (req, res, next) => {
+  try {
+    if (!req?.body?.userIds) {
+      return sendResponse(res, 400, 'User IDs required');
+    }
 
-    sendResponse(res, 200, 'Users status toggled successfully');
+    const { userIds } = req.body;
+
+    try {
+      // parallelize the asynchronous operations
+      await Promise.all(userIds.map((userId) => handleUserStatusUnblock(userId)));
+      return sendResponse(res, 200, 'Users status active success');
+    } catch (dbError) {
+      sendResponse(res, 400, String(dbError));
+    }
   } catch (err) {
     next(err);
   }
@@ -56,10 +80,23 @@ const updateUsers = async (req, res, next) => {
 
 const deleteUser = async (req, res, next) => {
   try {
-    const { userId } = req.params;
-    await User.findByIdAndDelete(userId);
+    if (!req?.params?.userId) {
+      return sendResponse(res, 400, 'User ID is required');
+    }
 
-    sendResponse(res, 200, 'User deleted successfully');
+    const { userId } = req.params;
+
+    try {
+      const user = await User.findByIdAndDelete(userId);
+
+      if (!user) {
+        return sendResponse(res, 404, `User ID ${userId} not found`);
+      }
+
+      return sendResponse(res, 200, 'User deleted successfully');
+    } catch (dbError) {
+      sendResponse(res, 400, String(dbError));
+    }
   } catch (err) {
     next(err);
   }
@@ -67,10 +104,14 @@ const deleteUser = async (req, res, next) => {
 
 const deleteUsers = async (req, res, next) => {
   try {
+    if (!req?.body?.userIds) {
+      return sendResponse(res, 400, 'User IDs required');
+    }
+
     const { userIds } = req.body;
     await User.deleteMany({ _id: { $in: userIds } });
 
-    sendResponse(res, 200, 'Users deleted successfully');
+    return sendResponse(res, 200, 'Users deleted successfully');
   } catch (err) {
     next(err);
   }
@@ -79,8 +120,8 @@ const deleteUsers = async (req, res, next) => {
 module.exports = {
   handleAdminPanel,
   getUsers,
-  updateUser,
-  updateUsers,
+  updateUsersBlock,
+  updateUsersUnblock,
   deleteUser,
   deleteUsers,
 };
